Changing StdoutPatterns to use pointers to Regex objects
Instead of list of strings, using list of pointers to Regex objects
is a cleaner and faster approach.
otherwise when matching exec contexts, we need to compile all the
stdout patterns to regex, and if a pattern is invalid we won't
know until the subprocess is done.

this approach ensures the expected stdout regex patterns are all
valid regex objects before running the subcommand, and there is no
need to compile the regex patterns later.