Implementing ConfFromEnvStr to read directly conf from envs
The cmd could be configured with environment variables which will
define a Default criteria to work with, instead of reading
the Conf from a file.

This feature needs to create the Conf from environment variable strings,
which is implementd by this function in this commmit.
The function does not use environment directly so it can be used in
other context and testing it will be easier as well.