Add test-build target to Makefile to run and test the built artifact
Basic tests of the built binary, to make sure it behaves as expected.
Tests are done with xecho, a bash script that supports different
exit codes and output messages, testing mute itself is easy.

I chose to use a bash script and the Makfile here because they're
portable enough and should be available on most systems, and won't
introduce extra dependency on the running host.