Makefile finds Debian package version from Debian changelog
The Debian package version for the source tar file was hard coded
in the Makefile and needed to be detected automatically.

Finding the version from the changelog is consistent with the rest of
creating the deb package, because the version from the changelog is going to
be used for the deb package metadata anyways.

This version value used in the Makefile was only used
to name the exported source tar file from the source tree
(used by the deb packaging process to build the source package).