Add RPM packaging, make pkg-rpm creates src and binary RPM packages
The RPM spec file currently is simple and uses the go tools installed
separately on the host, and requirements are not installed automatically
yet. Because currently many RPM based hosts provide older go versions,
including my dev machine. So building the RPM using mock and installing
go as a build requirement is tricky. For now we'll assume
required versions of go is intalled on the building host.