Makefile find mute version from source code not Git tags
Makefile relied on Git tags to find the latest version (which
is used to create packages). This was so that no package is created
without a tag, but when testing local packaging seems a bit cumbersome,
and could endup to be different thant the actual version.

Using the version constant of the program itself is more consistent.