Change mute.Exec() to be a method of new Target struct
Instead of passing many arguments to mute.Exec() restructure
the code to use the struct "Target" to define what command
to execute, when to mute (conf) and where to print the output,
and the size of the pre allocated buffer.

This breaks the API of mute package, but no change in the functionality
of the program.