Rename constants ENV_* to CamelCase as a more conventional format
The constants ENV_CONFIG, ENV_EXIT_CODES and ENV_STDOUT_PATTERN
specified the name of the environment variables that could be
used to configure mute.

The names were all caps (conventional for environment variables),
but such names for costants are not conventional for go projects,
so this commit renames those constants.

This commit changes the mute go package, but the behavior of the
program stays the same, as the name of the environment variables
are still in all caps (MUTE_CONF, ...), Just the name
constant names are changed.